+++ packages/jest-changed-files/src/hg.js
@@ -31,17 +31,19 @@ const ANCESTORS = [
 const adapter: SCMAdapter = {
   findChangedFiles: async (
     cwd: string,
+    roots: Array<Path>,
     options: Options,
   ): Promise<Array<Path>> =>
     new Promise((resolve, reject) => {
-      let args = ['status', '-amnu'];
+      const args = ['status', '-amnu'];
       if (options && options.withAncestor) {
         args.push('--rev', `ancestor(${ANCESTORS.join(', ')})`);
       } else if (options && options.changedSince) {
         args.push('--rev', `ancestor(., ${options.changedSince})`);
       } else if (options && options.lastCommit === true) {
-        args = ['tip', '--template', '{files%"{file}\n"}'];
+        args.push('-A');
       }
+      args.push(...roots);
       const child = childProcess.spawn('hg', args, {cwd, env});
       let stdout = '';
       let stderr = '';